+++ app/src/main/java/com/jin/jjinweather/feature/location/GeoPoint.kt
@@ -1,3 +1,3 @@
-package com.jin.jjinweather.layer.domain.model.location
+package com.jin.jjinweather.feature.location
 
 data class GeoPoint(val latitude: Double, val longitude: Double)
