+++ app/src/main/java/com/jin/jjinweather/feature/location/LocationRepository.kt
@@ -1,7 +1,5 @@
 package com.jin.jjinweather.feature.location
 
-import com.jin.jjinweather.layer.domain.model.location.GeoPoint
-
 interface LocationRepository {
     suspend fun currentGeoPoint(): GeoPoint
