+++ app/src/main/java/com/jin/jjinweather/feature/location/data/GeoPointDataSource.kt
@@ -1,6 +1,6 @@
 package com.jin.jjinweather.feature.location.data
 
-import com.jin.jjinweather.layer.domain.model.location.GeoPoint
+import com.jin.jjinweather.feature.location.GeoPoint
 
 interface GeoPointDataSource {
     fun currentGeoPoint(): Result<GeoPoint>
