+++ app/src/main/java/com/jin/jjinweather/feature/weather/WeatherRepository.kt
@@ -1,7 +1,6 @@
 package com.jin.jjinweather.feature.weather
 
 import com.jin.jjinweather.layer.domain.model.UiState
-import com.jin.jjinweather.layer.domain.model.weather.Weather
 
 interface WeatherRepository {
     suspend fun weatherAt(latitude: Double, longitude: Double): UiState<Weather>
