+++ app/src/main/java/com/jin/jjinweather/feature/weather/data/WeatherRepositoryImpl.kt
@@ -2,9 +2,9 @@ package com.jin.jjinweather.feature.weather.data
 
 import android.database.SQLException
 import com.jin.jjinweather.feature.location.data.GeoCodeDataSource
-import com.jin.jjinweather.layer.data.database.entity.WeatherEntity
+import com.jin.jjinweather.feature.weather.data.model.WeatherEntity
 import com.jin.jjinweather.layer.domain.model.UiState
-import com.jin.jjinweather.layer.domain.model.weather.Weather
+import com.jin.jjinweather.feature.weather.Weather
 import com.jin.jjinweather.feature.weather.WeatherRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
