+++ app/src/main/java/com/jin/jjinweather/feature/weather/data/model/WeatherEntity.kt
@@ -1,9 +1,9 @@
-package com.jin.jjinweather.layer.data.database.entity
+package com.jin.jjinweather.feature.weather.data.model
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.jin.jjinweather.layer.domain.model.weather.DailyWeather
-import com.jin.jjinweather.layer.domain.model.weather.HourlyWeather
+import com.jin.jjinweather.feature.weather.DailyWeather
+import com.jin.jjinweather.feature.weather.HourlyWeather
 
 /**
  * 위치 기반 날씨 정보
