+++ app/src/main/java/com/jin/jjinweather/layer/data/database/AppDatabase.kt
@@ -5,8 +5,9 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.jin.jjinweather.feature.location.data.GeoPointTrackingDataSource
 import com.jin.jjinweather.feature.weather.data.WeatherTrackingDataSource
-import com.jin.jjinweather.layer.data.database.entity.GeoPointEntity
-import com.jin.jjinweather.layer.data.database.entity.WeatherEntity
+import com.jin.jjinweather.feature.location.data.model.GeoPointEntity
+import com.jin.jjinweather.feature.weather.data.Converters
+import com.jin.jjinweather.feature.weather.data.model.WeatherEntity
 
 @Database(entities = [WeatherEntity::class, GeoPointEntity::class], version = 1)
 @TypeConverters(Converters::class)
