+++ app/src/main/java/com/jin/jjinweather/layer/domain/usecase/GetCurrentLocationWeatherUseCase.kt
@@ -1,7 +1,7 @@
 package com.jin.jjinweather.layer.domain.usecase
 
 import com.jin.jjinweather.layer.domain.model.UiState
-import com.jin.jjinweather.layer.domain.model.weather.Weather
+import com.jin.jjinweather.feature.weather.Weather
 import com.jin.jjinweather.feature.location.LocationRepository
 import com.jin.jjinweather.feature.weather.WeatherRepository
